removed svn inline tagging and renamed ping to example
