+++ build.xml
@@ -1,5 +1,4 @@
 <!-- ======================================================================= -->
-<!-- $Id: build.xml 13 2010-10-13 17:54:42Z dacracot $ -->
 <project name="tox" default="main" basedir="./">
 	<!-- =================================================================== -->
 	<!-- get enviroment from OS -->
