+++ src/gov/llnl/tox/api/authenticator.java
@@ -1,4 +1,4 @@
-package gov.llnl.restful.api;
+package gov.llnl.tox.api;
 //---------------------------------------------------
 import java.io.*;
 import java.net.*;
@@ -7,7 +7,7 @@ import javax.servlet.*;
 import javax.servlet.http.*;
 import org.apache.commons.codec.binary.Base64;
 //---------------------------------------------------
-import gov.llnl.restful.util.*;
+import gov.llnl.tox.util.*;
 //----------------------------------------------------
 public abstract class authenticator
 	{
