+++ src/gov/llnl/tox/toxServlet.java
@@ -7,8 +7,6 @@ import javax.servlet.http.*;
 //---------------------------------------------------
 import gov.llnl.tox.util.*;
 //----------------------------------------------------
-//-- $Id: toxServlet.java 5 2009-10-16 15:20:39Z dacracot $
-//----------------------------------------------------
 public class toxServlet extends HttpServlet
 	{
 	//-----------------------------------------------
