+++ src/gov/llnl/tox/util/debug.java
@@ -4,8 +4,6 @@ import java.io.*;
 import java.text.*;
 import javax.servlet.*;
 //----------------------------------------------------
-//-- $Id: debug.java 5 2009-10-16 15:20:39Z dacracot $
-//----------------------------------------------------
 public class debug
 	{
 	//-----------------------------------------------
