+++ src/gov/llnl/tox/util/ora.java
@@ -7,8 +7,6 @@ import javax.sql.*;
 //---------------------------------------------------
 import oracle.jdbc.*;
 //----------------------------------------------------
-//-- $Id: ora.java 5 2009-10-16 15:20:39Z dacracot $
-//----------------------------------------------------
 public class ora
 	{
 	private Connection conn;
