+++ src/sql/example.bdy.sql
@@ -2,12 +2,8 @@ set echo on
 spool ping.bdy.sql.err
 set define off
 
--- $Id: ping.bdy.sql 5 2009-10-16 15:20:39Z dacracot $
-
 CREATE OR REPLACE PACKAGE BODY ping
-	IS
-	/*------------------------------------------------------------------------*/
-	/*------------------------------------------------------------------------*/
+	AS
 	/*========================================================================*/
 	FUNCTION encode
 		(
