+++ src/sql/example.prv.sql
@@ -1,8 +1,6 @@
 set echo on
 spool ping.prv.sql.err
 
--- $Id: ping.prv.sql 5 2009-10-16 15:20:39Z dacracot $
-
 /*------------------------------------------------------------------------*/
 
 GRANT EXECUTE ON ping TO tox;
