+++ src/sql/example.usr.sql
@@ -1,13 +1,14 @@
 set echo on
 spool ping.usr.sql.err
 
--- $Id: ping.usr.sql 5 2009-10-16 15:20:39Z dacracot $
-
 /*------------------------------------------------------------------------*/
 
-CREATE USER example IDENTIFIED BY examplebaby DEFAULT TABLESPACE data TEMPORARY TABLESPACE temp;
+CREATE USER example IDENTIFIED BY examplebaby DEFAULT TABLESPACE users TEMPORARY TABLESPACE temp;
 
-GRANT connect, resource TO example;
+ALTER USER example QUOTA UNLIMITED ON users; 
+
+GRANT CONNECT TO example;
+GRANT RESOURCE TO example;
 
 /*------------------------------------------------------------------------*/
